tools: cpt2 -- Add a bunch of criu file stubs
To proceed even minimal restore we need a bunch of
criu files to be present on disk, even if they are
empty ones. So since this patch set aims to implement
conversion only a few entries (tasks, regular files and
memory) we generate stubs so crtools won't complain about
missing files.

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>