sockets: Separate socket buffer size setting into a helper
It is known that for unix sockets we need to setup
qlen first before restoring queue itself, otherwise
there might be not enough place for data.

Thus move the snippet into separate helper. Most
likely we will merge it back again once it's
been proved that for any kind of sockets it
is safe to setup all options before queue
data itself.

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>